refactor(Tags): convert to tailwind
