+++ src/components/Ui/Ui.tsx
@@ -1,7 +1,5 @@
 import { styled, media, css } from '@styles/styled';
-import { getTagUrl } from '@utils/url';
 import { motion } from 'framer-motion';
-import Link from 'next/link';
 import React from 'react';
 
 /* TODO: fix this type */
@@ -46,40 +44,6 @@ export const headingLinkStyle = css`
     display: block;
     visibility: hidden;
     transition: visibility 0.2s ease-in-out;
-  }
-`;
-
-const StyledTag = styled.li`
-  font-size: 14px;
-  font-family: ${({ theme }) => theme.font.contentSans};
-  border-radius: 2px;
-  text-align: center;
-  transition: font-weight 0.2s ease-in-out;
-
-  &:hover {
-    font-weight: bold;
-  }
-
-  > a {
-    display: inline-block;
-    margin-right: 16px;
-    text-decoration: underline;
-  }
-`;
-
-export const Tag: React.FC<{ tag: string; slug: string }> = ({ tag, slug }) => (
-  <StyledTag>
-    <Link href={getTagUrl(slug)}>
-      <a>#{tag}</a>
-    </Link>
-  </StyledTag>
-);
-export const Tags = styled.ul`
-  display: flex;
-  flex-wrap: wrap;
-
-  && {
-    padding: 16px 0;
   }
 `;
